create native tables on BQ for data and metadata
I ran into a lot of issues caused by the parquet file schema.
Most of the values for the demand time series were integers,
but there were some missing values. When saved to parquet
without specifying the schema, pandas/pyarrow will attempt
to infer the data types in the cols to create the schema. Seeing
mostly ints in the first few rows, it will assign iteger to the schema,
then down the line, there will be schema conflicts between "double" and
"int_64". To get around this, I defined the schema using pyarrow to
enforce the value column to be a float64.

I also had to delete any previous versions of parquet files in the data
lake and external tables in the data warehouse that were created with
the old schema and recreate them.